Removing humdrum files from Xcode and fixing build warning

+++ src/toolkit.cpp
@@ -1006,7 +1006,7 @@ void Toolkit::SetHumdrumBuffer(const char *data)
     }
 
 #else
-    int size = strlen(data) + 1;
+    int size = (int)strlen(data) + 1;
     m_humdrumBuffer = (char *)malloc(size);
     if (!m_humdrumBuffer) {
         // something went wrong
